requirements.txt
# Core ML/DL Frameworks
torch==2.1.0
tensorflow==2.15.0
tensorflow-recommenders==0.7.3
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0

# Feature Store
feast==0.35.0
feast[redis]==0.35.0

# Vector Databases & ANN
faiss-cpu==1.7.4
pymilvus==2.3.3
annoy==1.17.3

# Streaming & Message Queue
kafka-python==2.0.2
confluent-kafka==2.3.0
aiokafka==0.10.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
polars==0.19.19
pyarrow==14.0.1
dask==2023.11.0

# API & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Caching & Database
redis==5.0.1
pymongo==4.6.0
sqlalchemy==2.0.23
asyncpg==0.29.0

# Monitoring & Logging
prometheus-client==0.19.0
grafana-api==1.0.3
mlflow==2.8.1
wandb==0.16.0
loguru==0.7.2

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1
hydra-core==1.3.2
omegaconf==2.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
locust==2.17.0

# Model Serving
onnx==1.15.0
onnxruntime==1.16.3
tritonclient==2.62.0

# Deep Learning Extensions
deepctr-torch==0.2.9
transformers==4.35.2

# Utilities
click==8.1.7
tqdm==4.66.1
httpx==0.25.2
requests==2.31.0

# Data Validation
great-expectations==0.18.5
pydantic==2.5.0

# Experiment Tracking
optuna==3.4.0
ray[tune]==2.8.0
